requirements.txt
numpy
pandas
python-igraph
pyyaml
tqdm
autologging